refactor: delete avatars.ts, inline avatar fetch via public APIs
Move detectProviderFromUrl and extractHandleFromUrl into providers.ts.
Replace session-based avatar fetching with simple public URL approach
(GitHub .png, unavatar.io for X). Remove AvatarReauthModal and all
related plumbing from ProfileEditor.
