+++ ui/profile/ProfileLinkRow.tsx
@@ -3,7 +3,7 @@
 import CopyButton from "@/ui/common/buttons/CopyButton";
 import VerifiedBadge from "@/ui/profile/VerifiedBadge";
 import { extractDomain, isDiscordLink } from "@/lib/profile/profileLinks";
-import { detectProviderFromUrl } from "@/ui/links/avatars";
+import { detectProviderFromUrl } from "@/ui/links/providers";
 import type { ProfileLinkRowProps } from "./profileCardTypes";
 import { resolveIconSrc } from "./profileCardUtils";
